Use size_t to keep the value returned by strlen().
When using a variable to store the result of strlen(), make sure its
type is size_t just like the function returns, rather than using a
smaller type like int.

+++ src/input/input_dvb.c
@@ -1659,7 +1659,7 @@ static void render_text_area(osd_renderer_t* renderer, osd_object_t* osd, char* 
   /* The line to be printed next. */
   char text_line[512];
   int text_width, text_height;
-  int old_line_length, line_cursor;
+  size_t old_line_length, line_cursor;
   char* bound, *old_bound;
   
   *height = 0;
